pyproject.toml
[project]
name = "sgcn-sgc-prototype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.118.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
